requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.0
aiosignal==1.3.2
attrs==25.3.0
discord.py==2.5.2
frozenlist==1.6.0
idna==3.10
multidict==6.4.4
propcache==0.3.1
PyYAML==6.0.2
yarl==1.20.0
